bayesian_q_learning: added error to pacman movement and to pose messages and services. In proccess of making foods map probabilistic on chance of containing food. TODO: create new pacman probabilistic behaviors

+++ bayesian_q_learning/include/bayesian_q_learning/bayesian_game_state.h
@@ -14,8 +14,8 @@
 class BayesianGameState : public GameState
 {
   protected:
-    void observeGhost(int measurement_x_dist, int measurement_y_dist, int ghost_index);
-    void observePacman(int measurement_x, int measurement_y);
+    void observeGhost(double measurement_x_dist, double measurement_y_dist, int ghost_index);
+    void observePacman(double measurement_x, double measurement_y);
     bool observeAgent(pacman_msgs::AgentPoseService::Request &req, pacman_msgs::AgentPoseService::Response &res);
     bool is_finished_;
